pyproject.toml
[tool.poetry]
name = "geminipro-first"
version = "0.1.0"
description = ""
authors = ["Ryo Kuroyanagi <im.ryokuro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.3.1"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
